# Conflicts:
#	pyproject.toml
#	src/flag_gems/runtime/backend/_kunlunxin/ops/diff.py
#	src/flag_gems/runtime/backend/_kunlunxin/ops/softmax.py
#	src/flag_gems/runtime/backend/_kunlunxin/ops/square.py

+++ src/flag_gems/runtime/backend/_enflame/gcu300/ops/cat.py
@@ -1,3 +1,17 @@
+# Copyright 2026 FlagOS Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import logging
 from typing import List, Tuple, Union
 
@@ -51,10 +65,10 @@ def cat_copy_func_kernel_4(
     dim_size_in_d,
     dim_size_out,
     dim_prod_post,
-    dim_offset_a: tl.int64,
-    dim_offset_b: tl.int64,
-    dim_offset_c: tl.int64,
-    dim_offset_d: tl.int64,
+    dim_offset_a: tl.int32,
+    dim_offset_b: tl.int32,
+    dim_offset_c: tl.int32,
+    dim_offset_d: tl.int32,
     total_elements_a,
     total_elements_b,
     total_elements_c,
@@ -67,22 +81,22 @@ def cat_copy_func_kernel_4(
     if pid_y == 0:
         in_ptr = in_ptr_a
         dim_size_in = dim_size_in_a
-        dim_offset = tl.cast(dim_offset_a, tl.int64)
+        dim_offset = tl.cast(dim_offset_a, tl.int32)
         total_elements = total_elements_a
     elif pid_y == 1:
         in_ptr = in_ptr_b
         dim_size_in = dim_size_in_b
-        dim_offset = tl.cast(dim_offset_b, tl.int64)
+        dim_offset = tl.cast(dim_offset_b, tl.int32)
         total_elements = total_elements_b
     elif pid_y == 2:
         in_ptr = in_ptr_c
         dim_size_in = dim_size_in_c
-        dim_offset = tl.cast(dim_offset_c, tl.int64)
+        dim_offset = tl.cast(dim_offset_c, tl.int32)
         total_elements = total_elements_c
     else:
         in_ptr = in_ptr_d
         dim_size_in = dim_size_in_d
-        dim_offset = tl.cast(dim_offset_d, tl.int64)
+        dim_offset = tl.cast(dim_offset_d, tl.int32)
         total_elements = total_elements_d
 
     block_start = pid_x * BLOCK_X
